Improve PhoneNumberKit package
