--- Sources/PhoneNumberKit/Resources
@@ -1,1 +1,0 @@
-../../PhoneNumberKit/Resources
