+++ Sources/PhoneNumberKit/Resources/Original/PhoneNumberMetadata.xml
