+++ Sources/PhoneNumberKit/Resources/PhoneNumberMetadata.json
