+++ Sources/PhoneNumberKit/Resources/update_metadata.sh
