--- Sources/PhoneNumberKit/UI
@@ -1,1 +1,0 @@
-../../PhoneNumberKit/UI
